Move struct `HashValue`
